refactor(gui): remove renderer-local artifact/chat create action
The open-epic store's createArtifact action wrote artifact and chat
entries directly into the epic Y.Doc, but production creation is
host-RPC-only (epic.createArtifact / epic.createChat) - the sidebar
never wired the local action, and chat rows minted renderer-side would
bypass host-side setup entirely. Only tests used it, as a doc-seeding
utility.

Remove the action from the production store surface and move the
seeding into createArtifactInDocForTests in the test shims, transacting
with the store's (now exported) LOCAL_ORIGIN so projector and
render-count semantics stay identical. The optimistic rename/delete/
reparent actions remain - they are the deliberate fast path layered
under their host RPCs.
