+++ clients/gui-app/src/stores/epics/open-epic/__tests__/epic-projector.test.ts
@@ -6,6 +6,7 @@
 import "../../../../../__tests__/test-browser-apis";
 import { describe, expect, it } from "vitest";
 import * as Y from "yjs";
+import { createArtifactInDocForTests } from "./projection-helpers-test-shims";
 import {
   createOpenEpicStore,
   type EpicStreamClientFactory,
@@ -149,9 +150,9 @@ describe("epic-projector", () => {
     handle.dispose();
   });
 
-  it("createArtifact populates artifacts.byId, allIds and the tree", () => {
+  it("a seeded artifact populates artifacts.byId, allIds and the tree", () => {
     const { handle } = newSession();
-    const id = handle.store.getState().createArtifact("spec", null);
+    const id = createArtifactInDocForTests(handle.doc, "spec", null);
     const state = handle.store.getState();
     expect(state.artifacts.allIds).toContain(id);
     expect(state.artifacts.byId[id].kind).toBe("spec");
@@ -164,7 +165,7 @@ describe("epic-projector", () => {
 
   it("projects artifactRoomId metadata without changing tree display fields", () => {
     const { handle } = newSession();
-    const id = handle.store.getState().createArtifact("spec", null);
+    const id = createArtifactInDocForTests(handle.doc, "spec", null);
     const artifactBefore = handle.store.getState().artifacts.byId[id];
     const treeBefore = handle.store.getState().tree;
 
@@ -189,8 +190,8 @@ describe("epic-projector", () => {
 
   it("rename only re-allocates the renamed slot; siblings keep ===", () => {
     const { handle } = newSession();
-    const a = handle.store.getState().createArtifact("spec", null);
-    const b = handle.store.getState().createArtifact("ticket", null);
+    const a = createArtifactInDocForTests(handle.doc, "spec", null);
+    const b = createArtifactInDocForTests(handle.doc, "ticket", null);
     const beforeA = handle.store.getState().artifacts.byId[a];
     const beforeB = handle.store.getState().artifacts.byId[b];
 
@@ -206,7 +207,7 @@ describe("epic-projector", () => {
 
   it("title edit does not invalidate tree rootIds when set is unchanged", () => {
     const { handle } = newSession();
-    const a = handle.store.getState().createArtifact("spec", null);
+    const a = createArtifactInDocForTests(handle.doc, "spec", null);
     const treeBefore = handle.store.getState().tree;
 
     handle.store.getState().renameArtifact(a, "Renamed");
@@ -219,8 +220,8 @@ describe("epic-projector", () => {
 
   it("structural change (parent move) updates childrenByParent buckets", () => {
     const { handle } = newSession();
-    const parent = handle.store.getState().createArtifact("spec", null);
-    const child = handle.store.getState().createArtifact("ticket", null);
+    const parent = createArtifactInDocForTests(handle.doc, "spec", null);
+    const child = createArtifactInDocForTests(handle.doc, "ticket", null);
 
     handle.store.getState().reparentArtifact(child, parent);
 
@@ -233,10 +234,10 @@ describe("epic-projector", () => {
 
   it("delete preserves artifact child parent links for host cascade", () => {
     const { handle } = newSession();
-    const root = handle.store.getState().createArtifact("spec", null);
-    const mid = handle.store.getState().createArtifact("spec", root);
-    const leaf = handle.store.getState().createArtifact("ticket", mid);
-    const chat = handle.store.getState().createArtifact("chat", mid);
+    const root = createArtifactInDocForTests(handle.doc, "spec", null);
+    const mid = createArtifactInDocForTests(handle.doc, "spec", root);
+    const leaf = createArtifactInDocForTests(handle.doc, "ticket", mid);
+    const chat = createArtifactInDocForTests(handle.doc, "chat", mid);
 
     handle.store.getState().deleteArtifact(mid);
 
@@ -251,7 +252,7 @@ describe("epic-projector", () => {
 
   it("chat creation populates chats slice and tree as a chat node", () => {
     const { handle } = newSession();
-    const id = handle.store.getState().createArtifact("chat", null);
+    const id = createArtifactInDocForTests(handle.doc, "chat", null);
     const state = handle.store.getState();
     expect(state.chats.allIds).toContain(id);
     expect(state.chats.byId[id].title).toBe("New chat");
@@ -275,9 +276,9 @@ describe("epic-projector", () => {
 
   it("projected slices match a fresh full projection of the live Y.Doc", () => {
     const { handle } = newSession();
-    const a = handle.store.getState().createArtifact("spec", null);
-    const b = handle.store.getState().createArtifact("ticket", a);
-    const c = handle.store.getState().createArtifact("chat", null);
+    const a = createArtifactInDocForTests(handle.doc, "spec", null);
+    const b = createArtifactInDocForTests(handle.doc, "ticket", a);
+    const c = createArtifactInDocForTests(handle.doc, "chat", null);
     handle.store.getState().setEpicTitle("Parity Check");
     handle.store.getState().renameArtifact(b, "Ticket B");
     void a;
@@ -635,7 +636,7 @@ describe("epic-projector", () => {
 
   it("a tombstone-only change does not rebuild the live artifact tree or byId", () => {
     const { handle } = newSession();
-    const live = handle.store.getState().createArtifact("spec", null);
+    const live = createArtifactInDocForTests(handle.doc, "spec", null);
 
     const treeBefore = handle.store.getState().tree;
     const artifactsBefore = handle.store.getState().artifacts;
@@ -671,8 +672,8 @@ describe("epic-projector", () => {
 
   it("preserves live artifact rename identity semantics when tombstones exist", () => {
     const { handle } = newSession();
-    const a = handle.store.getState().createArtifact("spec", null);
-    const b = handle.store.getState().createArtifact("ticket", null);
+    const a = createArtifactInDocForTests(handle.doc, "spec", null);
+    const b = createArtifactInDocForTests(handle.doc, "ticket", null);
 
     handle.doc.transact(() => {
       const tombstones = new Y.Map<unknown>();
